requirements.txt
werkzeug==2.0.3
flask==2.0.3
flask-cors==3.0.10
python-dotenv
crewai
langchain
openai
runwayml